% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{cl}
\alias{cl}
\title{Sample a Chung-Lu graph}
\usage{
cl(theta, ...)
}
\arguments{
\item{theta}{TODO}
}
\value{
Never returns Poisson edges.
}
\description{
Sample a Chung-Lu graph
}
\examples{

# TODO

}
\seealso{
\code{\link[=fastRG]{fastRG()}}

Other bernoulli graphs: \code{\link{er}}
}
\concept{bernoulli graphs}
