% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di_dc_sbm.R
\name{di_dc_sbm}
\alias{di_dc_sbm}
\title{Sample a directed degree corrected stochastic blockmodel graph}
\usage{
di_dc_sbm(
  theta_in,
  theta_out,
  pi_in,
  pi_out,
  B,
  avg_deg = NULL,
  poisson_edges = TRUE,
  sort_nodes = FALSE,
  ...
)
}
\arguments{
\item{theta_in}{A vector specifying the degree distribution parameters
for node in-degree. Setting \code{theta = rep(1, n)} results in
a stochastic blockmodel without in-degree correction.}

\item{theta_out}{A vector specifying the degree distribution parameters
for node out-degree. Setting \code{theta = rep(1, n)} results in
a stochastic blockmodel without out-degree correction.}

\item{pi_in}{Block sampling proportions for incoming blocks. Must be
positive, but do not need to sum to one, as they will be
normalized internally. Note that \code{length(pi_in)} implicitly
specifies the number of incoming blocks. The number of
incoming and outgoing blocks need not be the same.}

\item{pi_out}{Block sampling proportions for outgoing blocks. Must be
positive, but do not need to sum to one, as they will be
normalized internally. Note that \code{length(pi_out)} implicitly
specifies the number of outgoing blocks. The number of
incoming and outgoing blocks need not be the same.}

\item{B}{A \code{length(pi_in)} by \code{length(pi_out)} matrix of block connection
probabilities. \code{B[i, j]} contains the probability that a node in
community \code{i} links to a node in community \code{j}. Does not need to
be symmetric.}

\item{avg_deg}{When specified, rescales parameter such that the
expected degree is \code{avg_deg} in the Poisson multi-graph. When
\code{poisson_edges = FALSE}, the resulting graph will have lower a
average degree than \code{avg_deg} due to lack of multi-edges. When
the graph is sparse, the expected number of edges for the Poisson
multi-graph and Bernoulli graph are nearly the
same. Defaults to \code{NULL}, such that no scaling occurs.}

\item{poisson_edges}{Logical indicating whether or not multi-edges are
allowed. Defaults to \code{TRUE}, which keeps multi-edges and produces
a multi-graph. When \code{FALSE}, only single edges are allowed, resulting
in a graph. See \strong{details} for some additional comments. Effected by
\code{simple} argument.}

\item{sort_nodes}{Logical indicating whether or not to sort the nodes
so that they are grouped by block. This incurs the the cost of a
radix sort, which is linear in the number of nodes in the graph.
Defaults to \code{FALSE}. Useful for plotting.}

\item{...}{
  Arguments passed on to \code{\link[=fastRG]{fastRG}}
  \describe{
    \item{\code{X}}{An \code{n} by \code{k_1} matrix.}
    \item{\code{S}}{A \code{k_1} by \code{k_2} matrix.}
    \item{\code{Y}}{A \code{d} by \code{k_2} matrix. Defaults to \code{X}.}
    \item{\code{simple}}{When \code{TRUE}, indicates that you want to work with undirected
graphs where self-loops and multi-edges are prohibited. Accomplishes
this by setting \code{directed = FALSE}, \code{allow_self_loops = FALSE}, and
\code{poisson_edges = FALSE}, and then ignoring arguments \code{directed},
\code{allow_self_loops} and \code{poisson_edges}. Defaults to \code{FALSE}.}
    \item{\code{directed}}{Logical indicating whether or not the graph should be
directed. Default is \code{directed = TRUE}. When \code{directed = FALSE},
symmetrizes \code{S} internally. When \code{X = Y} (which is the default when
no \code{Y} is specified), this results in a symmetric adjacency matrix
as output. When \code{avg_deg} is specified and the desired graph is directed,
the average degree scaling is on the out-degree of each node (or the
row sums if you prefer to think in terms of the adjacency matrix).
Effected by the \code{simple} argument.}
    \item{\code{allow_self_loops}}{Logical indicating whether edges are allowed from
a node back to itself. Defaults to \code{TRUE}. When \code{FALSE}, sampling
proceeds normally, and then self-loops are removed after sampling
is completed. Effected by the \code{simple} argument.}
    \item{\code{return_edge_list}}{Logical indicating whether to return an edgelist
rather than an adjacency matrix. Defaults to \code{FALSE}.}
  }}
}
\value{
A random Poisson (or Bernoulli) dot product graph. By default,
returns a \code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}} in CSC format (i.e. of abstract
class \code{CsparseMatrix}). When the graph is undirected, the \code{sparseMatrix}
will also be symmetric.

If \code{return_edge_list = TRUE}, instead returns the graph as an edgelist
in matrix format, with contains two columns called \code{from} and \code{to}.
The \code{from} column contains the index of the outgoing edge,
and the \code{to} column contains the index of the incoming edge.
}
\description{
Sample a directed degree corrected stochastic blockmodel graph
}
\details{
TODO: document the generative model

samples from a degree corrected stochastic blockmodel
pi a K vector, contains block sampling proportions
theta an n vector, contains degree parameters
B is K time K

Define \eqn{lambda_ij = theta[i] * theta[j]*B_{U,V}}
where U,V are blockmemberships of i and j, sampled from multinomial(pi)
i connects to j with probability 1- exp( -lambda_ij)

if lambda_ij is small, then 1- exp( -lambda_ij) approx lambda_ij
if avgDeg is set, then B is scaled so that the expected average
degree is avgDeg.

to make function easy to parameterize, this function does not allow
for different degree distributions between blocks.
}
\examples{

set.seed(27)

n <- 1000
k_in <- 10
k_out <- 3

B <- matrix(runif(k_in * k_out), nrow = k_in, ncol = k_out)

theta_in <- round(rlnorm(n, 2))   # in degree heterogeneity
theta_out <- round(rlnorm(n, 1))  # out degree heterogeneity

pi_in <- c(1, 1, 2, 3, 1, 3, 8, 2, 3, 1)
pi_out <- c(1, 2, 9)

A <- di_dc_sbm(theta_in, theta_out, pi_in, pi_out, B, avg_deg = 50)

params <- di_dc_sbm_params(theta_in, theta_out, pi_in, pi_out, B, avg_deg = 50)

expected(params$X, params$S, params$Y)

}
\seealso{
\code{\link[=fastRG]{fastRG()}}

Other stochastic block models: 
\code{\link{dc_mixed}()},
\code{\link{dc_overlapping}()},
\code{\link{dcsbm}()},
\code{\link{di_planted_partition_params}()},
\code{\link{di_planted_partition}()},
\code{\link{disbm_params}()},
\code{\link{disbm}()}

Other directed graphs: 
\code{\link{di_planted_partition}()},
\code{\link{disbm}()}
}
\concept{directed graphs}
\concept{stochastic block models}
