% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{er}
\alias{er}
\title{Sample an Erdos-Renyi graph}
\usage{
er(n, p = NULL, avgDeg = NULL, directed = FALSE, ...)
}
\arguments{
\item{n}{TODO}

\item{p}{TODO}

\item{avgDeg}{TODO}

\item{directed}{Defaults to \code{FALSE} for Erdos-Renyi graphs. The default
in the more general \code{\link[=fastRG]{fastRG()}} is \code{TRUE}.}
}
\value{
Never returns Poisson edges.
}
\description{
Sample an Erdos-Renyi graph
}
\examples{

# TODO

}
\seealso{
\code{\link[=fastRG]{fastRG()}}

Other bernoulli graphs: \code{\link{cl}}
}
\concept{bernoulli graphs}
