% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{expectation}
\alias{expectation}
\alias{expectation.undirected_factor_model}
\alias{expectation.directed_factor_model}
\title{Calculate the expected adjacency matrix}
\usage{
expectation(model, ...)

\method{expectation}{undirected_factor_model}(model, ...)

\method{expectation}{directed_factor_model}(model, ...)
}
\arguments{
\item{model}{A \code{\link[=directed_factor_model]{directed_factor_model()}} or an \code{\link[=undirected_factor_model]{undirected_factor_model()}}
object.}

\item{...}{Unused.}
}
\value{
The expected value of the adjacency matrix, conditional on the
latent factors \code{X} and \code{Y} (if the model is directed).
}
\description{
Calculate the expected adjacency matrix
}
