% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expected}
\alias{expected}
\title{Calculate the expected number of edges in Poisson RDPG graph}
\usage{
expected(X, S, Y = X)
}
\arguments{
\item{X}{An \code{n} by \code{k_1} matrix.}

\item{S}{A \code{k_1} by \code{k_2} matrix.}

\item{Y}{A \code{d} by \code{k_2} matrix. Defaults to \code{X}.}
}
\value{
A named list with three measurements from a Poisson RDPG(X, S, Y)
multi-graph:
\itemize{
\item \code{count}: Expected number of edges.
\item \code{degree}: Expected degree of each node.
\item \code{density}: Expected edge density. Equals the expected number of edges
divided by the total number of edges.
}
}
\description{
Calculate the expected number of edges in Poisson RDPG graph
}
\details{
Note that the running time of \code{\link[=fastRG]{fastRG()}} is proportional to
\code{count}, the expected number of edges in the graph.
}
\examples{

n <- 10000
d <- 1000

k1 <- 5
k2 <- 3

X <- matrix(rpois(n = n * k1, 1), nrow = n)
Y <- matrix(rpois(n = d * k2, 1), nrow = d)
S <- matrix(runif(n = k1 * k2, 0, .1), nrow = k1)

expected(X, S, Y)
}
