% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected.R
\name{expected_edges}
\alias{expected_edges}
\title{Calculate the expected number of edges in Poisson RDPG graph}
\usage{
expected_edges(factor_model, ...)
}
\arguments{
\item{factor_model}{TODO}

\item{...}{Ignored.}
}
\description{
underestimate when edge distribution is bernoulli
}
\details{
Note that the running time of \code{\link[=fastRG]{fastRG()}} is proportional to
\code{count}, the expected number of edges in the graph.
}
\examples{

n <- 10000
d <- 1000

k1 <- 5
k2 <- 3

X <- matrix(rpois(n = n * k1, 1), nrow = n)
Y <- matrix(rpois(n = d * k2, 1), nrow = d)
S <- matrix(runif(n = k1 * k2, 0, .1), nrow = k1)

expected(X, S, Y)
expected_degrees(X, S, Y)
expected_svds(X, S, Y)

}
