% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastRG-package.R
\docType{package}
\name{fastRG-package}
\alias{fastRG}
\alias{fastRG-package}
\title{fastRG: Sample Generalized Random Product Graphs in Linear Time}
\description{
Samples generalized random product graph, a
    generalization of a broad class of network models. Given matrices X,
    S, and Y with with non-negative entries, samples a matrix with
    expectation X S Y^T and independent Poisson or Bernoulli entries. The
    algorithm first samples the number of edges and then puts them down
    one-by-one.  As a result it is O(m) where m is the number of edges, a
    dramatic improvement over element-wise algorithms that which require
    O(n^2) operations to sample a random graph, where n is the number of
    nodes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexpghayes/fastRG}
  \item Report bugs at \url{https://github.com/alexpghayes/fastRG/issues}
}

}
\author{
\strong{Maintainer}: Alex Hayes \email{alexpghayes@gmail.com} (0000-0002-4985-5160)

Authors:
\itemize{
  \item Karl Rohe \email{KarlRohe@stat.wisc.edu}
  \item Jun Tao
  \item Xintian Han
  \item Norbert Binkiewicz
}

}
\keyword{internal}
