% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{howManyEdges}
\alias{howManyEdges}
\title{Calculate the expected number of edges in Poisson gRDPG graph}
\usage{
howManyEdges(X, S, Y = X)
}
\arguments{
\item{X}{An \code{n} by \code{k_1} matrix.}

\item{S}{A \code{k_1} by \code{k_2} matrix.}

\item{Y}{A \code{d} by \code{k_2} matrix. Defaults to \code{X}.}
}
\value{
A named list with three measurements from a Poisson gRDPG(X, S, Y)
multi-graph:
\itemize{
\item \code{em}: Expected number of edges.
\item \code{avgDeg}: Expected degree of each node.
\item \code{density}: Expected edge density. Equals the expected number of edges
divided by the total number of edges.
}
}
\description{
Calculate the expected number of edges in Poisson gRDPG graph
}
\details{
Note that the running time of \code{\link[=fastRG]{fastRG()}} is proportional to the
expected number of edges \code{em}.
}
\examples{

# TODO

}
