% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undirected_overlapping_sbm.R
\name{overlapping_sbm}
\alias{overlapping_sbm}
\title{Create an undirected overlapping degree corrected stochastic blockmodel object}
\usage{
overlapping_sbm(
  n = NULL,
  theta = NULL,
  k = NULL,
  B = NULL,
  ...,
  pi = rep(1/k, k),
  sort_nodes = TRUE
)
}
\arguments{
\item{n}{(degree heterogeneity) The number of nodes in the blockmodel.
Use when you don't want to specify the degree-heterogeneity
parameters \code{theta} by hand. When \code{n} is specified, \code{theta}
is randomly generated from a \code{LogNormal(2, 1)} distribution.
This is subject to change, and may not be reproducible.
\code{n} defaults to \code{NULL}. You must specify either \code{n}
or \code{theta}, but not both.}

\item{theta}{(degree heterogeneity) A numeric vector
explicitly specifying the degree heterogeneity
parameters. This implicitly determines the number of nodes
in the resulting graph, i.e. it will have \code{length(theta)} nodes.
Must be positive. Setting to a vector of ones recovers
a stochastic blockmodel without degree correction.
Defaults to \code{NULL}. You must specify either \code{n} or \code{theta},
but not both.}

\item{k}{(mixing matrix) The number of blocks in the blockmodel.
Use when you don't want to specify the mixing-matrix by hand.
When \code{k} is specified, the elements of \code{B} are drawn
randomly from a \code{Uniform(0, 1)} distribution.
This is subject to change, and may not be reproducible.
\code{k} defaults to \code{NULL}. You must specify either \code{k}
or \code{B}, but not both.}

\item{B}{(mixing matrix) A \code{k} by \code{k} matrix of block connection
probabilities. The probability that a node in block \code{i} connects
to a node in community \code{j} is \code{Poisson(B[i, j])}. Must be
a square matrix. \code{matrix} and \code{Matrix} objects are both
acceptable. If \code{B} is not symmetric, it will be
symmetrized via the update \code{B := B + t(B)}. Defaults to \code{NULL}.
You must specify either \code{k} or \code{B}, but not both.}

\item{...}{
  Arguments passed on to \code{\link[=undirected_factor_model]{undirected_factor_model}}
  \describe{
    \item{\code{expected_degree}}{If specified, the desired expected degree
of the graph. Specifying \code{expected_degree} simply rescales \code{S}
to achieve this. Defaults to \code{NULL}. Do not specify both
\code{expected_degree} and \code{expected_density} at the same time.}
    \item{\code{expected_density}}{If specified, the desired expected density
of the graph. Specifying \code{expected_density} simply rescales \code{S}
to achieve this. Defaults to \code{NULL}. Do not specify both
\code{expected_degree} and \code{expected_density} at the same time.}
  }}

\item{pi}{(relative block probabilities) Relative block
probabilities. Must be positive, but do not need to sum
to one, as they will be normalized internally.
Must match the dimensions of \code{B} or \code{k}. Defaults to
\code{rep(1 / k, k)}, or a balanced blocks.}

\item{sort_nodes}{Logical indicating whether or not to sort the nodes
so that they are grouped by block and by \code{theta}. Useful for plotting.
Defaults to \code{TRUE}.}
}
\value{
An \code{undirected_overlapping_sbm} S3 object, a subclass of the
\code{\link[=undirected_factor_model]{undirected_factor_model()}} with the following additional
fields:
\itemize{
\item \code{Z}: The community memberships of each node, as an \code{n} by \code{k}
binary indicator matrix.
\item \code{theta}: A numeric vector of degree-heterogeneity parameters.
\item \code{pi}: Sampling probabilities for each block.
\item \code{sorted}: Logical indicating where nodes are arranged by
block (and additionally by degree heterogeneity parameter)
within each block.
}
}
\description{
To specify a degree-corrected stochastic blockmodel, you must specify
the degree-heterogeneity parameters (via \code{n} or \code{theta}),
the mixing matrix (via \code{k} or \code{B}), and the relative block
probabilites (optional, via \code{pi}). We provide sane defaults for most of these
options to enable rapid exploration, or you can invest the effort
for more control over the model parameters. We \strong{strongly recommend}
setting the \code{expected_degree} or \code{expected_density} argument
to avoid large memory allocations associated with
sampling large, dense graphs.
}
\section{Generative Model}{
Note: we implement a slight generalization of the original overlapping
stochastic blockmodel introduce in Zhang et al. (2014) called the
Overlapping Continuous Community Assignment Model (OCCAM).

There are two levels of randomness in a degree-corrected
overlapping stochastic blockmodel. First, for each node, we
independently determine if that node is a member of each block. This is
handled by \code{overlapping_sbm()}. Then, given these block memberships,
we randomly sample edges between nodes. This second
operation is handled by \code{\link[=sample_edgelist]{sample_edgelist()}},
\code{\link[=sample_sparse]{sample_sparse()}}, \code{\link[=sample_igraph]{sample_igraph()}} and
\code{\link[=sample_tidygraph]{sample_tidygraph()}}, depending depending on your desired
graph representation.
\subsection{Block memberships}{

Note that some nodes may not belong to any blocks.

\strong{TODO}
}

\subsection{Degree heterogeneity}{

In addition to block membership, the overlapping SBM also allows
nodes to have different propensities for edge formation.
We represent this propensity for node \eqn{i} by a positive
number \eqn{\theta_i}.

\strong{TODO} identifiability constraint?
}

\subsection{Edge formulation}{

Once we know the block memberships and the degree
heterogeneity parameters \eqn{theta}, we need one more
ingredient, which is the baseline intensity of connections
between nodes in block \code{i} and block \code{j}. Then each edge
\eqn{A_{i,j}} is Poisson distributed with parameter

\strong{TODO}
}
}

\examples{

set.seed(27)

lazy_overlapping_sbm <- overlapping_sbm(n = 1000, k = 5, expected_density = 0.01)
lazy_overlapping_sbm

# sometimes you gotta let the world burn and
# sample a wildly dense graph

dense_lazy_overlapping_sbm <- overlapping_sbm(n = 500, k = 3, expected_density = 0.8)
dense_lazy_overlapping_sbm

# explicitly setting the degree heterogeneity parameter,
# mixing matrix, and relative community sizes rather
# than using randomly generated defaults

k <- 5
n <- 1000
B <- matrix(stats::runif(k * k), nrow = k, ncol = k)

theta <- round(stats::rlnorm(n, 2))

pi <- c(1, 2, 4, 1, 1) / 5

custom_overlapping_sbm <- overlapping_sbm(
  theta = theta,
  B = B,
  pi = pi,
  expected_degree = 50
)

custom_overlapping_sbm

edgelist <- sample_edgelist(custom_overlapping_sbm)
edgelist

# efficient eigendecompostion that leverages low-rank structure in
# E(A) so that you don't have to form E(A) to find eigenvectors,
# as E(A) is typically dense. computation is
# handled via RSpectra

population_eigs <- eigs_sym(custom_overlapping_sbm)

}
\references{
Kaufmann, Emilie, Thomas Bonald, and Marc Lelarge.
"A Spectral Algorithm with Additive Clustering for the Recovery of
Overlapping Communities in Networks," Vol. 9925.
Lecture Notes in Computer Science.
Cham: Springer International Publishing, 2016.
https://doi.org/10.1007/978-3-319-46379-7.

Latouche, Pierre, Etienne Birmelé, and Christophe Ambroise.
"Overlapping Stochastic Block Models with Application to the
French Political Blogosphere." The Annals of Applied Statistics 5,
no. 1 (March 2011): 309–36. https://doi.org/10.1214/10-AOAS382.

Zhang, Yuan, Elizaveta Levina, and Ji Zhu. "Detecting
Overlapping Communities in Networks Using Spectral Methods."
ArXiv:1412.3432, December 10, 2014. http://arxiv.org/abs/1412.3432.
}
\seealso{
Other stochastic block models: 
\code{\link{dcsbm}()},
\code{\link{directed_dcsbm}()},
\code{\link{planted_partition}()},
\code{\link{sbm}()}

Other undirected graphs: 
\code{\link{chung_lu}()},
\code{\link{dcsbm}()},
\code{\link{erdos_renyi}()},
\code{\link{planted_partition}()},
\code{\link{sbm}()}
}
\concept{stochastic block models}
\concept{undirected graphs}
