% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{sbm}
\alias{sbm}
\title{Sample a stochastic blockmodel graph}
\usage{
sbm(n, pi, B, avgDeg = NULL, PoissonEdges = TRUE, ...)
}
\arguments{
\item{n}{TODO}

\item{pi}{Block sampling proportions.}

\item{B}{"Contains probabilities". TODO}

\item{avgDeg}{When specified, the expected average degree of nodes
in the output poisson multi-graph. When \code{poissonEdges = FALSE}, the
resulting graph will have lower average degree due to lack
of multiple-edges. When the graph is sparse, the expected number of
edges for the Poisson multi-graph and Bernoulli graph are nearly the
same. Defaults to \code{NULL}, such that no scaling occurs.}

\item{PoissonEdges}{Logical indicating whether or not multi-edges are
allowed. Defaults to \code{TRUE}, which keeps multi-edges and produces
a multi-graph. When \code{FALSE}, only single edges are allowed, resulting
in a graph. See \strong{details} for some additional comments. Effected by
\code{simple} argument.}
}
\value{
A random Poisson (or Bernoulli) dot product graph. By default,
returns a \code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}} in CSC format (i.e. of abstract
class \code{CsparseMatrix}). When the graph is undirected, the \code{sparseMatrix}
will also be symmetric.

If \code{returnEdgeList = TRUE}, instead returns the graph as an edgelist
in matrix format, where the first column of the matrix contains two
columns \code{from} and \code{to} . The \code{from} column contains the index
of the outgoing edge in this case, and the \code{to} column contains
the index of the incoming edge.
}
\description{
Sample a stochastic blockmodel graph
}
\details{
TODO
}
\examples{

# TODO

}
\seealso{
\code{\link[=fastRG]{fastRG()}}

Other stochastic block models: \code{\link{dcMixed}},
  \code{\link{dcOverlapping}}, \code{\link{dcsbm}}
}
\concept{stochastic block models}
