% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected-spectra.R
\name{svds.undirected_factor_model}
\alias{svds.undirected_factor_model}
\title{Compute the singular value decomposition of the expected adjacency matrix of an undirected factor model}
\usage{
\method{svds}{undirected_factor_model}(A, k = A$k, nu = k, nv = k, opts = list(), ...)
}
\arguments{
\item{A}{An \code{\link[=undirected_factor_model]{undirected_factor_model()}}.}

\item{k}{Desired rank of decomposition.}

\item{nu}{Number of left singular vectors to be computed. This must
be between 0 and \code{k}.}

\item{nv}{Number of right singular vectors to be computed. This must
be between 0 and \code{k}.}

\item{opts}{Control parameters related to the computing
algorithm. See \strong{Details} below.}

\item{...}{Unused, included only for consistency with generic signature.}
}
\description{
Compute the singular value decomposition of the expected adjacency matrix of an undirected factor model
}
