% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directed_erdos_renyi.R
\name{directed_erdos_renyi}
\alias{directed_erdos_renyi}
\title{Create an directed erdos renyi object}
\usage{
directed_erdos_renyi(n, ..., p = NULL)
}
\arguments{
\item{n}{Number of nodes in graph.}

\item{...}{
  Arguments passed on to \code{\link[=directed_factor_model]{directed_factor_model}}
  \describe{
    \item{\code{X}}{A \code{\link[=matrix]{matrix()}} or \link[=matrix]{Matrix()} representing real-valued
latent node positions encoding community structure of
incoming edges. Entries must be positive.}
    \item{\code{S}}{A \code{\link[=matrix]{matrix()}} or \link[=matrix]{Matrix()} mixing matrix. Entries
must be positive.}
    \item{\code{Y}}{A \code{\link[=matrix]{matrix()}} or \link[=matrix]{Matrix()} representing real-valued
latent node positions encoding community structure of
outgoing edges. Entries must be positive.}
    \item{\code{expected_in_degree}}{If specified, the desired expected degree
of the graph. Specifying \code{expected_degree} simply rescales \code{S}
to achieve this. Defaults to \code{NULL}. Do not specify both
\code{expected_degree} and \code{expected_density} at the same time. TODO}
    \item{\code{expected_out_degree}}{If specified, the desired expected degree
of the graph. Specifying \code{expected_degree} simply rescales \code{S}
to achieve this. Defaults to \code{NULL}. Do not specify both
\code{expected_degree} and \code{expected_density} at the same time. TODO}
    \item{\code{expected_density}}{If specified, the desired expected density
of the graph. Specifying \code{expected_density} simply rescales \code{S}
to achieve this. Defaults to \code{NULL}. Do not specify both
\code{expected_degree} and \code{expected_density} at the same time. TODO}
  }}

\item{p}{Probability of an edge between any two nodes. You must specify
either \code{p}, \code{expected_in_degree}, or \code{expected_out_degree}.}
}
\value{
Never returns Poisson edges.
}
\description{
Create an directed erdos renyi object
}
\examples{

set.seed(87)

er <- directed_erdos_renyi(n = 10, p = 0.1)
er

big_er <- directed_erdos_renyi(n = 10^6, expected_in_degree = 5)
big_er

A <- sample_sparse(er)
A

}
\seealso{
Other bernoulli graphs: 
\code{\link{erdos_renyi}()}

Other erdos renyi: 
\code{\link{erdos_renyi}()}
}
\concept{bernoulli graphs}
\concept{directed graphs}
\concept{erdos renyi}
