% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected-spectra.R
\name{eigs_sym.undirected_factor_model}
\alias{eigs_sym.undirected_factor_model}
\title{Compute the eigendecomposition of the expected adjacency matrix of an undirected factor model}
\usage{
\method{eigs_sym}{undirected_factor_model}(A, k = A$k, which = "LM", sigma = NULL, opts = list(), ...)
}
\arguments{
\item{A}{An \code{\link[=undirected_factor_model]{undirected_factor_model()}}.}

\item{k}{Desired rank of decomposition.}

\item{which}{Selection criterion. See \strong{Details} below.}

\item{sigma}{Shift parameter. See section \strong{Shift-And-Invert Mode}.}

\item{opts}{Control parameters related to the computing
algorithm. See \strong{Details} below.}

\item{...}{Unused, included only for consistency with generic signature.}
}
\description{
Compute the eigendecomposition of the expected adjacency matrix of an undirected factor model
}
