% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erdos_renyi.R
\name{erdos_renyi}
\alias{erdos_renyi}
\title{Create an undirected erdos renyi object}
\usage{
erdos_renyi(n, ..., p = NULL)
}
\arguments{
\item{n}{Number of nodes in graph.}

\item{p}{Probability of an edge between any two nodes. You must specify
either \code{p} or \code{expected_degree}.}
}
\value{
Never returns Poisson edges.
}
\description{
Create an undirected erdos renyi object
}
\examples{

set.seed(87)

er <- erdos_renyi(n = 10, p = 0.1)
er

big_er <- erdos_renyi(n = 10^6, expected_degree = 5)
big_er

A <- sample_sparse(er)
A

}
\seealso{
\code{\link[=fastRG]{fastRG()}}

Other bernoulli graphs: 
\code{\link{directed_erdos_renyi}()}

Other erdos renyi: 
\code{\link{directed_erdos_renyi}()}

Other undirected graphs: 
\code{\link{dcsbm}()},
\code{\link{planted_partition}()},
\code{\link{sbm}()}
}
\concept{bernoulli graphs}
\concept{erdos renyi}
\concept{undirected graphs}
