% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directed_erdos_renyi.R, R/erdos_renyi.R
\name{erdos_renyi}
\alias{erdos_renyi}
\title{Create an directed erdos renyi object}
\usage{
erdos_renyi(n, ..., p = NULL)

erdos_renyi(n, ..., p = NULL)
}
\arguments{
\item{n}{Number of nodes in graph.}

\item{p}{Probability of an edge between any two nodes. You must specify
either \code{p} or \code{expected_degree}.}
}
\value{
Never returns Poisson edges.

Never returns Poisson edges.
}
\description{
Create an directed erdos renyi object

Create an undirected erdos renyi object
}
\examples{

set.seed(87)

er <- directed_erdos_renyi(n = 10, p = 0.1)
er

big_er <- directed_erdos_renyi(n = 10^6, expected_in_degree = 5)
big_er

A <- sample_sparse(er)
A


set.seed(87)

er <- erdos_renyi(n = 10, p = 0.1)
er

big_er <- erdos_renyi(n = 10^6, expected_degree = 5)
big_er

A <- sample_sparse(er)
A

}
\seealso{
\code{\link[=fastRG]{fastRG()}}

\code{\link[=fastRG]{fastRG()}}

Other bernoulli graphs: 
\code{\link{chung_lu}()},
\code{\link{di_planted_partition}()}

Other directed graphs: 
\code{\link{di_dc_sbm}()},
\code{\link{di_planted_partition}()},
\code{\link{disbm}()}

Other bernoulli graphs: 
\code{\link{chung_lu}()},
\code{\link{di_planted_partition}()}

Other undirected graphs: 
\code{\link{dcsbm}()},
\code{\link{di_planted_partition_params}()},
\code{\link{sbm}()}
}
\concept{bernoulli graphs}
\concept{directed graphs}
\concept{erdos renyi}
\concept{undirected graphs}
