% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erdos_renyi.R
\name{erdos_renyi}
\alias{erdos_renyi}
\alias{erdos_renyi_params}
\title{Sample an Erdos-Renyi graph}
\usage{
erdos_renyi(n, p = NULL, avg_deg = NULL, directed = FALSE, ...)

erdos_renyi_params(n, p = NULL, avg_deg = NULL, directed = FALSE, ...)
}
\arguments{
\item{n}{Number of nodes in graph.}

\item{p}{Probability of an edge between any two nodes. You must specify
either \code{p} or \code{avg_deg}. If you do not specify \code{p}, uses \code{avg_deg / n}.}

\item{avg_deg}{Desired average out degree. When specified, rescales
sampling probabilities to achieve the desired average out degree.
Defaults to \code{NULL}, such that there is no rescaling.}

\item{directed}{Defaults to \code{FALSE} for Erdos-Renyi graphs. The default
in the more general \code{\link[=fastRG]{fastRG()}} is \code{TRUE}.}

\item{...}{
  Arguments passed on to \code{\link[=fastRG]{fastRG}}
  \describe{
    \item{\code{X}}{An \code{n} by \code{k_1} matrix.}
    \item{\code{S}}{A \code{k_1} by \code{k_2} matrix.}
    \item{\code{Y}}{A \code{d} by \code{k_2} matrix. Defaults to \code{X}.}
    \item{\code{simple}}{When \code{TRUE}, indicates that you want to work with undirected
graphs where self-loops and multi-edges are prohibited. Accomplishes
this by setting \code{directed = FALSE}, \code{allow_self_loops = FALSE}, and
\code{poisson_edges = FALSE}, and then ignoring arguments \code{directed},
\code{allow_self_loops} and \code{poisson_edges}. Defaults to \code{FALSE}.}
    \item{\code{poisson_edges}}{Logical indicating whether or not multi-edges are
allowed. Defaults to \code{TRUE}, which keeps multi-edges and produces
a multi-graph. When \code{FALSE}, only single edges are allowed, resulting
in a graph. See \strong{details} for some additional comments. Effected by
\code{simple} argument.}
    \item{\code{allow_self_loops}}{Logical indicating whether edges are allowed from
a node back to itself. Defaults to \code{TRUE}. When \code{FALSE}, sampling
proceeds normally, and then self-loops are removed after sampling
is completed. Effected by the \code{simple} argument.}
    \item{\code{return_edge_list}}{Logical indicating whether to return an edgelist
rather than an adjacency matrix. Defaults to \code{FALSE}.}
  }}
}
\value{
Never returns Poisson edges.
}
\description{
Sample an Erdos-Renyi graph
}
\examples{

set.seed(27)

# get the random dot product parameters
erdos_renyi_params(n = 10, p = 0.1)

# sample a small graph
A <- erdos_renyi(n = 10, p = 0.1)
A

# out degree of each node
rowSums(A)

# in degree of each node
colSums(A)

# sample a much larger graph
B <- erdos_renyi(n = 10^6, avg_deg = 5)

}
\seealso{
\code{\link[=fastRG]{fastRG()}}

Other bernoulli graphs: 
\code{\link{chung_lu}()},
\code{\link{di_planted_partition_params}()},
\code{\link{di_planted_partition}()}
}
\concept{bernoulli graphs}
