% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{eigs_sym}
\alias{svds}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{RSpectra}{\code{\link[RSpectra:eigs]{eigs_sym}}, \code{\link[RSpectra]{svds}}}
}}

